dvr-scan-py/pyproject.toml
[project]
name = "dvr-scan-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyewf-ctypes",
]
